refactor(runtime): extract active compact facts
